+++ packages/bundler/src/cli/commands/run/batcher.ts
@@ -2,10 +2,15 @@ import { Command } from "commander";
 import { BundlerBatcher } from "../../../batcher";
 import { makeLogger } from "@nocturne-xyz/offchain-utils";
 import { getRedis } from "./utils";
+import { extractConfigName } from "@nocturne-xyz/config";
 
 const runBatcher = new Command("batcher")
   .summary("run bundler batcher")
   .description("must supply .env file with REDIS_URL and REDIS_PASSWORD.")
+  .requiredOption(
+    "--config-name-or-path <string>",
+    "config name or path to Nocturne contract JSON config file"
+  )
   .option("--batch-size <number>", "batch size")
   .option(
     "--max-latency <number>",
@@ -21,9 +26,17 @@ const runBatcher = new Command("batcher")
     "min log importance to log to stdout. if not given, logs will not be emitted to stdout"
   )
   .action(async (options) => {
-    const { maxLatency, batchSize, logDir, stdoutLogLevel } = options;
+    const { configNameOrPath, maxLatency, batchSize, logDir, stdoutLogLevel } =
+      options;
 
-    const logger = makeLogger(logDir, "bundler", "batcher", stdoutLogLevel);
+    // TODO: consolidate batcher and submitter into one component (batcher doesn't need config name)
+    const configName = extractConfigName(configNameOrPath);
+    const logger = makeLogger(
+      logDir,
+      `${configName}-bundler`,
+      "batcher",
+      stdoutLogLevel
+    );
     const batcher = new BundlerBatcher(
       getRedis(),
       logger,