+++ packages/bundler/src/cli/commands/run/server.ts
@@ -3,7 +3,7 @@ import { ethers } from "ethers";
 import { BundlerServer } from "../../../server";
 import { makeLogger } from "@nocturne-xyz/offchain-utils";
 import { getRedis } from "./utils";
-import { loadNocturneConfig } from "@nocturne-xyz/config";
+import { extractConfigName, loadNocturneConfig } from "@nocturne-xyz/config";
 
 const runServer = new Command("server")
   .summary("run bundler server")
@@ -34,7 +34,13 @@ const runServer = new Command("server")
     }
     const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
 
-    const logger = makeLogger(logDir, "bundler", "server", stdoutLogLevel);
+    const configName = extractConfigName(configNameOrPath);
+    const logger = makeLogger(
+      logDir,
+      `${configName}-bundler`,
+      "server",
+      stdoutLogLevel
+    );
     const server = new BundlerServer(
       config.tellerAddress(),
       provider,