+++ packages/bundler/src/cli/commands/run/submitter.ts
@@ -3,7 +3,11 @@ import { ethers } from "ethers";
 import { BundlerSubmitter } from "../../../submitter";
 import { makeLogger } from "@nocturne-xyz/offchain-utils";
 import { getRedis } from "./utils";
-import { loadNocturneConfig } from "@nocturne-xyz/config";
+import { extractConfigName, loadNocturneConfig } from "@nocturne-xyz/config";
+import {
+  DefenderRelayProvider,
+  DefenderRelaySigner,
+} from "@openzeppelin/defender-relay-client/lib/ethers";
 
 const runSubmitter = new Command("submitter")
   .summary("run bundler submitter")
@@ -27,22 +31,41 @@ const runSubmitter = new Command("submitter")
     const { configNameOrPath, logDir, stdoutLogLevel } = options;
     const config = loadNocturneConfig(configNameOrPath);
 
+    const relayerApiKey = process.env.OZ_RELAYER_API_KEY;
+    const relayerApiSecret = process.env.OZ_RELAYER_API_SECRET;
+
     const privateKey = process.env.TX_SIGNER_KEY;
-    if (!privateKey) {
-      throw new Error("missing TX_SIGNER_KEY");
-    }
-
     const rpcUrl = process.env.RPC_URL;
-    if (!rpcUrl) {
-      throw new Error("missing RPC_URL");
-    }
-    const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
-    const signingProvider = new ethers.Wallet(privateKey, provider);
 
-    const logger = makeLogger(logDir, "bundler", "submitter", stdoutLogLevel);
+    let signer: ethers.Signer;
+    if (relayerApiKey && relayerApiSecret) {
+      const credentials = {
+        apiKey: relayerApiKey,
+        apiSecret: relayerApiSecret,
+      };
+      const provider = new DefenderRelayProvider(credentials);
+      signer = new DefenderRelaySigner(credentials, provider, {
+        speed: "average",
+      });
+    } else if (rpcUrl && privateKey) {
+      const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
+      signer = new ethers.Wallet(privateKey, provider);
+    } else {
+      throw new Error(
+        "missing RPC_URL/PRIVATE_KEY or OZ_RELAYER_API_KEY/OZ_RELAYER_API_SECRET"
+      );
+    }
+
+    const configName = extractConfigName(configNameOrPath);
+    const logger = makeLogger(
+      logDir,
+      `${configName}-bundler`,
+      "submitter",
+      stdoutLogLevel
+    );
     const submitter = new BundlerSubmitter(
       config.tellerAddress(),
-      signingProvider,
+      signer,
       getRedis(),
       logger
     );