+++ packages/e2e-tests/tests/optimisticRecordTracking.ts
@@ -179,8 +179,14 @@ describe("Optimistic nullifier tracking", () => {
     const op = await proveOperation(joinSplitProver, signedOp);
     await submitAndProcessOperation(op);
 
-    // ensure it removes all records when it polls op digest from bundler
+    // ensure it removes all records when it polls op digest from bundler (mock date to bypass
+    // update op digest buffer)
+    const dateNow = Date.now;
+    Date.now = () => {
+      return dateNow() + 60 * 1000;
+    };
     await sdk.updateOptimisticNullifiers();
+    Date.now = dateNow;
 
     // DB should have no more optimistic records
     const nfRecordsAfter = await db.getAllOptimisticNFRecords();