+++ packages/sdk/src/proveOperation.ts
@@ -4,47 +4,37 @@ import {
   SignedOperation,
   ProvenJoinSplit,
   ProvenOperation,
+  toSubmittableOperation,
+  SubmittableOperationWithNetworkInfo,
 } from "./primitives";
 import {
   JoinSplitProver,
   joinSplitPublicSignalsFromArray,
   packToSolidityProof,
 } from "./proof";
+import { iterChunks } from "./utils";
+
+// SDK will fire off at most this many provers in parallel
+const MAX_PARALLEL_PROVERS = 4;
 
 export async function proveOperation(
   prover: JoinSplitProver,
   op: SignedOperation
-): Promise<ProvenOperation> {
-  const joinSplits: ProvenJoinSplit[] = await Promise.all(
-    op.joinSplits.map((joinSplit) => proveJoinSplit(prover, joinSplit))
-  );
+): Promise<SubmittableOperationWithNetworkInfo> {
+  const joinSplits: ProvenJoinSplit[] = [];
+  for (const batch of iterChunks(op.joinSplits, MAX_PARALLEL_PROVERS)) {
+    const provenBatch = await Promise.all(
+      batch.map((joinSplit) => proveJoinSplit(prover, joinSplit))
+    );
+    joinSplits.push(...provenBatch);
+  }
 
-  const {
-    networkInfo,
-    refundAddr,
-    encodedRefundAssets,
-    actions,
-    encodedGasAsset,
-    gasAssetRefundThreshold,
-    executionGasLimit,
-    gasPrice,
-    deadline,
-    atomicActions,
-  } = op;
-
-  return {
-    networkInfo,
+  const operation: ProvenOperation = {
+    ...op,
     joinSplits,
-    refundAddr,
-    encodedRefundAssets,
-    actions,
-    encodedGasAsset,
-    gasAssetRefundThreshold,
-    executionGasLimit,
-    gasPrice,
-    deadline,
-    atomicActions,
   };
+
+  return toSubmittableOperation(operation);
 }
 
 async function proveJoinSplit(