+++ packages/deposit-screener/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- replace tx manager with OZ relay
 - revert back to ethers eip712 signer after fixing typehash ordering
 - switch on screening/delay components in server/processor CLI
 - delete all dummy magic values and hardcode into dummy impls