+++ packages/frontend-sdk/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+- `sync` method takes and propogates `SyncOpts` as params to snap/sdk
+- instantiate fe-sdk with `config`
+- add method to fetch operation status from bundler
 - in `verifyProvenOperation`, mask `encodedAsset` PIs to 0 if `publicSpend` is 0
 - forward `SubmittableOperation` to bundler now that operation format has changed post min return value changes
 - instantiate fe-sdk with just `configName`