+++ packages/snap/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+- Add `lastSyncedMerkleIndex` usage if sync already is taking place
+- remove custom gas est call in snap and let sdk handle
+- `nocturne_sync` takes and propagates `SyncOpts` as params
 - change spend key derivation to `keccak256(bip44Node.privateKey)`
 - remove use of winston logger
 - fix bug where snap was parsing op metadata as `.metadata` not `.opMetadata`