+++ packages/test-actor/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- replace tx manager with OZ relay
 - use .network instead of .chainid for op request building
 - add CLI option to perform 8 ops in rapid succession to fill a bundle every K ops
 - allow setting separate intervals for deposits and ops in CLI